requirements.txt
flask
pypdf
sentence-transformers
faiss-cpu
numpy
nltk
